pyproject.toml
[tool.poetry]
name = "mpulse-challenge"
version = "0.1.0"
description = ""
authors = ["Jeff Skelton <jeff@jeffskelton.net>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.82.0"
uvicorn = "^0.18.3"
pydantic = "^1.10.1"
aiokafka = "^0.7.2"
SQLAlchemy = "^1.4.40"
psycopg2-binary = "^2.9.3"
alembic = "^1.8.1"
python-multipart = "^0.0.5"
SQLAlchemy-Utils = "^0.38.3"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
python-dotenv = "^0.21.0"
pytest-postgresql = "^4.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
